Now, MappedFieldQueryFilter uses excludecurator
MappedFieldQueryFilter pays attention to an excludecurator argument
in the Query. MappedFieldQueryFilter is the class we use that
interprets custom fields and ties them to Lucene columns.

In our case, we actually tie to them to N Lucene columns -- one per
available provenance. We nicknamed this "column explosion" in Lucene.

But in the case that the user passes in excludecurator:XYZ, we look
up all the Feeds owned by that curator, and exclude data from those
Feeds.