Log message cleanup in del.icio.us plugin.
A VERY unlikely situation where attempting to compile an invalid
XPathExpr will throw an exception triggering a log. We have statically
defined expressions in our code, so the chance of them being invalid
is nil. So we should never see this log message.

Additionally, Eclipse was having issues with formatting cleanups due
to the length of the string in the previous log message, this one
makes it much happier and the auto-formatted statement is actually
easier to read.