Removing TestMinusCurator
This test was supposed to be an end-to-end test that added some
feeds curated by different parties and made sure that when you
search for them using the Nutch web UI, we could exclude
curators.

However, the test calls "ant" and takes other actions that mutate
the state of the working tree, and it does not know how to clean
up after itself. I am throwing away the test "for now," hoping that
one day we can have an integration test framework that is more
sensible.