Move ProvenancePredicatePair testing to TestRdfStoreFactory
This test method used to be in its own test, TestPPP, but the
testing that we do is more than a ProvenancePredicatePair unit test --

it is a full test of the ability of RdfStoreFactory to generate
ProvenancePredicatePair objects and of the working of the ProvenancePredicatePair
object.

Moving it here also highlights the fact that the only reason the
ProvenancePredicatePair exists is to provide a String that represents
the smushing of the provenance and predicate together. I wonder if that
would make more sense to move straight into the TripleStoreIndexer, and
then we could throw away ProvenancePredicatePairs entirely.