In MappedFieldQueryFilter, use existing logic to get list of excluded curators
In this commit, we make the getCuratorsToExclude() method in DocumentExclusionBasedOnCuratorQueryFilter
public and static. This permits MappedFieldQueryFilter to use that method to loop across the input
clauses and get the list of curators that the user passed in to excludecurator.

We re-use code here rather than re-implement because we have found this to be error-prone.