Move record extraction into crawl phase
Before this commit, each class that implemented IResourceExtractor
had to request the OAI:PMH Resource from the OAI:PMH server.

After this commit, the interface has changed: the process() method
takes not the server but the actual Resource object. That way, the
process() method does not have to go out to the server.

It also means that the OAIPMH importer in our code can use
listRecords() instead of getting each record one at a time. This should
improve the speed of the aggregate step.