Provide a crawl-and-merge.sh script
This is DiscoverEd-specific, since it uses hard-coded paths for the crawl
path and the seed directory of URLs.

However, it makes things extraordinarily simpler, so I am committing it. It
is possible to override its assumptions through command line options; however,
they are not documented except in the script source.